my: improve performance
Use grep to filter out files not in home directory, as that is much faster
than doing it with locate.

As a result, don’t pass --regex to locate, as it’s not needed: the user can
decide whether or not to pass it.

Also, don’t pass --existing to locate, as that should be up to the user (and
I have a front-end to locate that does that anyway).

Note that the output is coloured if to a tty.